refactor: storeHero에서 reviewInfo 분리
